chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/parsing/subtree-visibility-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Subtree Visibility Test: subtree-visibility with valid values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<meta name="assert" content="subtree-visibility supports the full grammar 'visible | auto | hidden | hidden-matchable'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("subtree-visibility", "visible");
test_valid_value("subtree-visibility", "auto");
test_valid_value("subtree-visibility", "hidden");
test_valid_value("subtree-visibility", "hidden-matchable");
</script>
</body>
</html>
